requirements.txt
torch>=2.0.0
transformers>=4.30.0
librosa>=0.10.0
soundfile>=0.12.0
openai-whisper>=20231117
anthropic>=0.18.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
